+++ routes/steamless.ts
@@ -1,9 +1,11 @@
+import { downloadFromUrl } from "../util/downloadFromUrl";
 import express from "express";
 import fetch from "cross-fetch";
 import crypto from "crypto"
 import path from "path"
 import fs from "fs"
-import { downloadFromUrl } from "../util/downloadFromUrl";
+import hashTable from "../hashTable.json"
+import * as jsonUtil from "../util/jsonUtil"
 
 export const steamlessRouter = express.Router()
 steamlessRouter.get("/", async (req, res) => {
@@ -28,18 +30,32 @@ steamlessRouter.get("/", async (req, res) => {
             const name = Match[4]
             const version = Match[2]
             const hash = crypto.createHash('sha256');
-
             let fileHash;
-            try {
-                await downloadFromUrl(url, "tmp")
-                const file = fs.readFileSync(`tmp/${fileName}`)
-                hash.update(file);
-                fileHash = hash.digest("hex")
-                fs.unlink(`tmp/${fileName}`, (err) => {
-                    if (err) throw err
-                })
-            } catch (error) {
-                console.log(error);
+
+            const jsonData = fs.readFileSync("./hashTable.json", "utf-8")
+            const data = JSON.parse(jsonData)
+
+            if (data.steamless.version == version) {
+                fileHash = data.steamless.hash
+            } else {
+                try {
+                    await downloadFromUrl(url, "tmp")
+                    const file = fs.readFileSync(`tmp/${fileName}`)
+                    hash.update(file);
+                    fileHash = hash.digest("hex")
+                    fs.unlink(`tmp/${fileName}`, (err) => {
+                        if (err) throw err
+                    })
+                    data.steamless.hash = fileHash
+                    data.steamless.version = version
+                    fs.writeFile("./hashTable.json", JSON.stringify(data), err => {
+                        if (err) {
+                            console.log(err)
+                        }
+                    })
+                } catch (error) {
+                    console.log(error);
+                }
             }
 
             const fullUrl: string = `${req.protocol}://${req.get('host')}${req.originalUrl}`
